+++ user-rest-service/domain/model/user.go
