+++ user-rest-service/domain/repository/repository.go
@@ -4,6 +4,10 @@ import (
 	"github.com/paypay3/kakeibo-app-api/user-rest-service/domain/model"
 )
 
+type DBRepository interface {
+	UserRepository
+}
+
 type UserRepository interface {
 	FindID(signUpUser *model.SignUpUser) error
 	FindEmail(signUpUser *model.SignUpUser) error
